Modify the button counter's count display element
Modify so it is easier to display the user's current count. Switch it from a <p> element into a span so it is displayed through javascript easier. Button attributes were edited. Remove the class label on the buttons because it is redundent with a the clasification by Id. Background image file path  wrong and website not looking good.